chnages

+++ app/Http/Controllers/Front/AllController.php
@@ -1,11 +1,14 @@
 <?php
+
 namespace App\Http\Controllers\Front;
+
 use App\Http\Controllers\Controller;
 use App\Models\OnlineUsers;
 use App\Models\Rooms;
 use App\Models\UserList;
 use Illuminate\Http\Request;
 use Nette\Utils\Random;
+
 class AllController extends Controller
 {
     public function welcome()
@@ -65,16 +68,18 @@ class AllController extends Controller
             return response()->json(["res" => "success", "user" => $user]);
         }
     }
-    public function skipping()
+    public function skipping(Request $request)
     {
         $ip = $_SERVER['REMOTE_ADDR'];
         $user =  UserList::where('ip_address', $ip)->first();
         UserList::where('id', $user->id)->update([
             "status" => 0
         ]);
-        $old_room = Rooms::where('my_id',$user->id)->orwhere('other_id',$user->id)->first();
-        if(!empty($old_room)){
-            Rooms::destroy($old_room->id);
+        if ($request->myid != 0 && $request->otherid != 0) {
+            $old_room = Rooms::where('my_id', $request->myid)->orwhere('other_id', $request->otherid)->first();
+            if (!empty($old_room)) {
+                Rooms::destroy($old_room->id);
+            }
         }
         $allusers = null;
         if (!empty($user->connect_with)) {
@@ -102,7 +107,9 @@ class AllController extends Controller
                 ->orwhere('status', 2)
                 ->get();
         }
+
         foreach ($allusers as $alluser) {
+
             if (empty(OnlineUsers::where('user_id', $alluser->id)->first())) {
                 OnlineUsers::create([
                     'user_id' => $alluser->id
@@ -110,44 +117,46 @@ class AllController extends Controller
             }
         }
         $date = date("Y-m-d");
-        $room_name = rand(99999999,10000000);
+        $room_name = rand(99999999, 10000000);
         while (true) {
-            $online_users = $this->all_online_user();
+            $online_users = $this->all_online_user($user->id);
             if (count($online_users) >= 2) {
-                $user1 = $online_users[0];
-                $user2 = $online_users[1];
+                $other = 0;
+                $user1 = $online_users[0]["user_id"];
+                $user2 = $online_users[1]["user_id"];
                 if ($user1 == $user->id || $user2 == $user2) {
-                    $room = Rooms::where('room_date',$date)->where('my_id',$user->id)->orwhere('other_id',$user->id)->first();
-                    UserList::where('id',$user1)->update([
-                        "status"=> 1
+                    $room = Rooms::where('room_date', $date)->where('my_id', $user->id)->orwhere('other_id', $user->id)->first();
+                    UserList::where('id', $user1)->update([
+                        "status" => 1
                     ]);
-                    UserList::where('id',$user2)->update([
-                        "status"=> 1
+                    UserList::where('id', $user2)->update([
+                        "status" => 1
                     ]);
 
-                    if(!empty($room)){
-                        return response()->json(["res"=>"success","room"=>$room]);
-                    }else{
+                    if (!empty($room)) {
+                        return response()->json(["res" => "success", "room" => $room]);
+                    } else {
                         Rooms::create([
-                           'my_id'=> $user1,
-                           'other_id' => $user2,
-                           'room_date'=>$date,
-                           'room_name' => $room_name
+                            'my_id' => $user1,
+                            'other_id' => $user2,
+                            'room_date' => $date,
+                            'room_name' => $room_name
                         ]);
-                        $room = Rooms::where('room_date',$date)->where('my_id',$user->id)->orwhere('other_id',$user->id)->first();
-                        return response()->json(["res"=>"success","room"=>$room]);
+                        $room = Rooms::where('room_date', $date)->where('my_id', $user->id)->orwhere('other_id', $user->id)->first();
+
+                        return response()->json(["res" => "success", "room" => $room]);
                     }
-                    OnlineUsers::where('user_id',$user1)->delete();
-                    OnlineUsers::where('user_id',$user2)->delete();
+                    OnlineUsers::where('user_id', $user1)->delete();
+                    OnlineUsers::where('user_id', $user2)->delete();
                     exit();
                 }
             }
             sleep(1);
         }
     }
-    public function all_online_user()
+    public function all_online_user($id)
     {
-        $allonline_users = OnlineUsers::pluck('user_id')->toArray();
+        $allonline_users = OnlineUsers::inRandomOrder()->limit(2)->get()->toArray();
         return $allonline_users;
     }
 }
